added timers module, refactor
